nghttpx: Extend Downstream stream_id to 64 bits

+++ src/shrpx_downstream.h
@@ -319,20 +319,20 @@ enum class DispatchState {
 
 class Downstream {
 public:
-  Downstream(Upstream *upstream, MemchunkPool *mcpool, int32_t stream_id);
+  Downstream(Upstream *upstream, MemchunkPool *mcpool, int64_t stream_id);
   ~Downstream();
   void reset_upstream(Upstream *upstream);
   Upstream *get_upstream() const;
-  void set_stream_id(int32_t stream_id);
-  int32_t get_stream_id() const;
-  void set_assoc_stream_id(int32_t stream_id);
-  int32_t get_assoc_stream_id() const;
+  void set_stream_id(int64_t stream_id);
+  int64_t get_stream_id() const;
+  void set_assoc_stream_id(int64_t stream_id);
+  int64_t get_assoc_stream_id() const;
   void pause_read(IOCtrlReason reason);
   int resume_read(IOCtrlReason reason, size_t consumed);
   void force_resume_read();
   // Set stream ID for downstream HTTP2 connection.
-  void set_downstream_stream_id(int32_t stream_id);
-  int32_t get_downstream_stream_id() const;
+  void set_downstream_stream_id(int64_t stream_id);
+  int64_t get_downstream_stream_id() const;
 
   int attach_downstream_connection(std::unique_ptr<DownstreamConnection> dconn);
   void detach_downstream_connection();
@@ -566,12 +566,12 @@ private:
   // How many times we tried in backend connection
   size_t num_retry_;
   // The stream ID in frontend connection
-  int32_t stream_id_;
+  int64_t stream_id_;
   // The associated stream ID in frontend connection if this is pushed
   // stream.
-  int32_t assoc_stream_id_;
+  int64_t assoc_stream_id_;
   // stream ID in backend connection
-  int32_t downstream_stream_id_;
+  int64_t downstream_stream_id_;
   // RST_STREAM error_code from downstream HTTP2 connection
   uint32_t response_rst_stream_error_code_;
   // An affinity cookie value.
